Allow to create a change for the initial commit of refs/meta/config
Normally you can only push commits for review to branches that exist.
But there is an exception for the branch to which HEAD in the remote
repository points (usually master). The same is true on submit, changes
can only be submitted if the destination branch exists, except if the
destination branch is the branch to which HEAD points. Relax this
further and also allow to push an initial commit for review to the
refs/meta/config branch and also allow to submit such a change.

This is useful to setup an initial project configuration when the
refs/meta/config branch is missing.

Actually already today it is possible to create a change with an initial
commit for the refs/meta/config branch, but then this change is not
submittable because the destination branch is missing. The creation of
such a change is possible by editing the access rights in the WebUI and
then clicking on the 'Save for Review' button. This creates a change for
the refs/meta/config branch regardless of whether the branch exists.

Change-Id: I2e28fdd5384149d0ee32b10641652803c73842ab
Signed-off-by: Edwin Kempin <ekempin@google.com>

Former-commit-id: f5a46a41096da6b3af81ad6719309b08ac3702a5