Show related changes when conflicts do not resolve
The server is known to throw 400 errors when it finds too many
conflicts. The client expects the resolved promise to return an array.
Because the template still tries to render the array (even though it is
not visible), having _conflicts be undefined causes rendering to break.

This change adds a check for an undefined API response and sets _changes
to be an empty array in that case.

Bug: Issue 6981
Change-Id: Ifee888a3ced02b0d9383cc9110b062b9cbc57848

Former-commit-id: a7016e39ae53e81446eb9ad26f2545ca08b3de09