Simplify artifacts deployment to Maven Central
De-duplicate the pom.xml definition in top-level directories and rename
the single left pom.xml file as fake_pom.xml. Roll-back the process of
version update in pom.xml file and remove the version.py utility. It's
not needed, as the version is passed to the mvn program on the command
line.

This is the preparation change for flattening the directory structure
and dissolving gerrit-* top-level directories.

Change-Id: Ied4d580e5ce2e98858f1f8e248969da1cb7d3a9a

Former-commit-id: 0ce2fdc1c22701745b8c3e1bb9b3f49de6b34de0