Hide patchset controls when edit is loaded
The patch set description and diff preferences controls are hidden when
an edit is loaded. Descriptions are not supported for edits, and diff
prefs are hidden to make room for edit specific controls to come in a
later change.

Also colors the patchset header the same shade as the WIP mode header
when an edit is loaded -- this involved a small refactor to some
computed class names.

Bug: Issue 4437
Change-Id: I2cc242e174b25403ed5ad8586c8dffb59fbd933d

Former-commit-id: dc41a315c426dd3e5de21dc1b2f5e0033cd60328