Correct group lookup by name from the index
The case sensitive name was used when adding a group to the index but
the lower case version was used for lookups. As this doesn't work for
names containing capital letters, we switch to a case sensitive lookup.
(Lookups for the full name of an account also use a case sensitive
lookup.)

Change-Id: I90bcd7d4b0918031f0d7431841eabb57e2fdfde8

Former-commit-id: 43d5115c98d4c188d6606644ef51ea163accf4d3