AccountApi: Add getGroups method
Add a method to get the groups to which an account belongs.

Note that the test user for some reason does not get included into
the "Anonymous Users" and "Registered Users" groups, while the admin
user does. Fixing this is outside the scope of this commit so it's
left as-is with a TODO to update the test later.

Bug: Issue 7065
Change-Id: If4c2d878adb90b020d74369316953796dbfcb712

Former-commit-id: b151e1eb805b8307aea6a5a17488311fa6654aa5