Implement 'Pure Revert' REST endpoint
In some cases, users want to find out if a change is a pure revert of
some other change. A change is a pure revert of another change, if it
just reverts the change and performs no additional content
modifications.

This commit adds a new REST endpoint that can be called on a change
resource and given an optional commit SHA1 as query parameter to check
against. If the parameter is omitted, the change is checked
against the reference of 'revertOf' if it exists.

Change-Id: Ia2ce051afe7d24aed688bfa3d8bbd1cb8fe4a200

Former-commit-id: bb84fd71c9ed81900c9f3731f701a4bb44c86149