Support labels >33 chars in LabelVote#parseWithEquals
In 57ecf24c back in 2013, I accidentally wrote this:
  Short.parseShort(text.substring(e + 1), text.length())
where *obviously* I meant this:
  Short.parseShort(text.substring(e + 1, text.length()))

This called Short#parseShort(String, int), where the second arg is a
radix. (The second arg to substring is also not necessary in this case,
but would have been harmlessly correct if it were present.)

It's surprising and lucky that passing an arbitrary radix managed to
have correct behavior in practically all cases, but this was in fact the
case due a combination of two facts:
  * text.length() is always at least 3 ("X=0"), usually longer, and
    only exceeds Character.MAX_RADIX (36) when the label name exceeds 33
    characters.
  * The value itself is rarely greater than 2 or 3, essentially never
    requiring multiple digits (unless someone were foolhardy enough to
    use a very large range), so any radix greater than or equal to the
    value would result in parsing the value correctly.

As an amusing side effect, the old code would parse "Code-Review=c"
"correctly," but would fail on "Code-Review=d".

Fix this bug and increase test coverage.

Change-Id: I17f4fb963e1e0d7b7bd51d689e7557cbcfc8b474

Former-commit-id: 0085f48710cdcf13baec5dd7b48b138776efac29