OnlineReindexer: Include index name in task name and log messages
The online reindexer was originally written to reindex changes, but has
since been extended to also support reindexing accounts and groups. In
future it will also support reindexing other index types, for example
projects.

The reindexer task name only includes the schema versions, but not the
index name, so it's not possible to know for which index the online
reindex task is executing. Also none of the log messages include the
index name, except the error message on reindex failure, but that is
hard-coded to say "changes".

Update the task name and log messages to show the correct index name.

Change-Id: I14e06c3db908cbda2812a5684823b3d1d300084d

Former-commit-id: 361755fcaa33f2b2133261271cc625643594972f