Use ~ instead of - as delimiter for local comments
Projects can contain dashes as part of their name which confuses the
local comments storage. By using tilde instead, we can work around this
issue because it is part of the invalid strings collection that prevents
creation of new projects containing any of those strings.

Once deployed, we start neglecting local comments that use the old
delimiter. Since comments are synced to the server immediately if the
user is online, there should be almost no cases where comments are lost.
In addition, a comment would still be in the browser's local storage and
retrievable from there manually.

Change-Id: Ibb18635e359a3b338b04e6a0e95637f699d2ebef

Former-commit-id: 1bd33e313201517bbc70e4d24d9a690595eabff1