Update bower2bazel to fetch correct packages
Previously, bower2bazel tried to fetch whatever was registered in
bower with the package name. If the package author was specified it was
ignored. There was an issue with installing a new package (paper-button)
in which bower tried to fetch the wrong dependency. (paper-ripple).

This change updates the package to use the original source, so that the
bower-archives file looks like:

bower_archive(
    name = "paper-ripple",
    package = "polymerelements/paper-ripple",
    version = "1.0.10",
    sha1 = "21199db50d02b842da54bd6f4f1d1b10b474e893")

Change-Id: I4d1f797a86bd80e8b9cf119e21ee7c2f5387b77e

Former-commit-id: 8a70455ae8c1175d5c021920636439f2923020a7