Add tests to make sure ssh commands can be executed
Recently a commit was submitted that caused the execution of some ssh
commands to fail with internal server error, and this was not noticed
during review/test because we don't have test coverage of all the ssh
commands.

Add a basic test that executes each of the ssh commands with the --help
option. This of course won't test all of the commands' functionality,
but at least will prevent any future breakages such as mentioned above.

Change-Id: I69a88e593018d5838bd06943feb8df7992819f0b

Former-commit-id: d9804401b6106818e113cd501ac450b61d7b56fe