ReceiveCommits: Update caller fullName after everything else
We have a useful feature of populating the fullName field in the Account
for a caller based on a corresponding identity in the commits that they
push. However, this was awkwardly wedged into the middle of the loop
where commits are validated, hurting readability. It is also a side
effect that is not otherwise taken into account when determining whether
the operation succeeds, so a push that apparently fails might actually
have this observable side effect.

Change the loop to just keep track of the full name it discovers, and
perform this side effect after all other work.

Change-Id: I0ae4b5628d483ba2b842173b9dc8e10804edeb98

Former-commit-id: 1ea05bdc99d4b25f652b6471a7c70344237c06d6