Rename bypass-review to skip-validation and document it
This push option is misnamed; It does not control whether the user is
allowed to bypass review. By definition any push directly to a branch
(rather than refs/for/) is bypassing review, which is the wording that
is used in user-upload.txt and elsewhere in the documentation.

Rename the option to describe what it actually does: it skips the commit
validation step. Describe this behavior in the push documentation.

Since this is now properly documented, make the constant public as well.

Change-Id: I012e1ea42a9a92a07c09d32beb9471a55e84508c

Former-commit-id: bf46140d7720132c3cf20f4515606fdc1eb110bb