Use the download attribute on patch links
Patch links were being wrongfully captured by the catchall route and
resulted in a visual 404 when clicked. The "pass-through" approach to
instructing page.js to ignore these links does not work because clicking
the link should result in a download and should not change the browser's
URL.

However, page.js will ignore links with the download attribute, and this
perfectly aligns with the intent of the problematic links. With this
change all download links in the download dialog bear the `download`
attribute and a test ensures it.

Tests are lightly cleaned up to avoid repeating some setup code.

Bug: Issue 7149
Change-Id: I83fa1b4bbd0d3b6d347d0b057ca7033db8cd82d6

Former-commit-id: ebf009cc5c7ee4e35392baed5922caafeb6ee89c