Revert "Ensure undefined patch ranges are always populated"
This reverts commit 5fbc5d932c8fc01e1d43b450e32557478c6ab4c3 [formerly e441e485161ebea3c9926d21a342b38a93e1789f].

The change this reverts sets undefined patch ranges in more cases but
does so using information that may be out of date in some of them.
Specifically, when navigating between changes, the `_allPatchSets` list
is derived from the change being navigated away from rather than the new
change being loaded.

The underlying cause of issue 7027 is instead fixed in I9d16616ccf.

Change-Id: I0f8855a0cb94116d339f5c2a2b17ddf2bb84f737

Former-commit-id: e438c21d8df27678f761c89a07ccb39e813c9bed