Introduce Change Report formatter extension point
When change is pushed for review typical response looks like:
  remote: New Changes:
  remote:   http://[server]/gerrit/[change-id] [change subject]

It is not sufficient for sites that serve Gerrit as a service that is
integrated under unified UI (with unified Code Browser). In such
environment links to Gerrit are no longer valid and it should be
possible to alter them with desired format.

This patch introduces ChangeReportFormatter that can be replaced with
instance loaded from plugin (DynamicItem pattern) with default
implementation that preserves current behavior.

Change-Id: Ia2de44c121b6a73fcf93e1aaf11690697adb9abf
Signed-off-by: Jacek Centkowski <jcentkowski@collab.net>
Signed-off-by: David Pursehouse <dpursehouse@collab.net>

Former-commit-id: 1b27d904dee1d64e042fd4cbb160830aa9940b68