Refactor file list header into change view
A future iteration of PolyGerrit will combine the contents that are
above the filelist and in the current filelist header. This change
refactors the change view and filelist so that all of that content is
part of the change view. This will enable moving elements around in a
single HTML element.

Change-Id: Iadb939643087589e16d7d90e77fde8f62a457f0a

Former-commit-id: 6805fd149041a322e3e63312abb2ac93e8a2ad06