LabelNormalizer: stop squashing based on permitted range
We are leveraging on this class to do votes squashing. Currently, The
following squashings will happen after changing permissions:

1. The label definition is removed. This will result in all votes of
this label to be deleted.
2. The label definition is changed. Existing votes will be squashed
to the left/right bound if they exceed.
3. User's permission range is updated. Existing votes of this user
will be squashed to the left/right permitted bound if they exceed.
4. User's permission range is delete. Existing votes will be deleted.

After this change, for case 3&4, votes will not be squashed or deleted
any more. If some votes are kept over time, the user's permission at
the voting moment will be honored.

Change-Id: I85c3e10c5b710b248e70fee2dd434ca9cecf9764

Former-commit-id: ff5fbbfc08a08b1ff9c910cac26df6204a658602