OnlineNoteDbMigrationIT: Use local disk
Even though the migration is online so we only need to start a server
once, part of the migration process includes side effects of writing new
migration state to config files on disk.

Somewhat confusingly, the online migrator was actually writing contents
to a gerrit.config file on disk, even though that file was not where the
config used by the in-memory server was stored.

Change-Id: I0ba86907e01cd63eb006f876911e1d7d9a6eccef

Former-commit-id: 4002210eab7162ded8e58f8bed7ddda009bcb4e8