Handle improperly encoded `+` in URLs
The `+` character is canonically used in URLs to indicate an encoded
space. Thus, in some instances, a PolyGerrit link with a legitimate `+`
(e.g. in between project and change num) may have that `+` encoded to
`%20`, which decodes to a space.

This change adds another route that catches this specific case and
redirects.

Also fixes a typo in gr-router_test.html.

Bug: Issue 6888, Issue 7100
Change-Id: Iadd1e63e726c5a7b1fd81afe3553556ddfb8012f

Former-commit-id: 5484d86f273966541de5146d74de06991aee9b16