Do cleanup when change action requests fail
The progress indicator when executing a change action needs to get
"cleaned up" when the action is complete. For example, when
cherry-picking a change, the progress indicator says "Cherry picking..."
and the message disappears when the cherry pick change has been
successfully created.

However, when the request fails, for example when a change cannot be
cherry-picked for git tree reasons, resulting in an HTTP 409 and an
error message, the progress indicator still needs to be cleaned up.

With this change, cleanup is fired from the change action _send method's
failure modes, and tests are added to encode these modes.

Bug: Issue 7184
Change-Id: If69f73a09a39a6dd94b838409aaeddf25f001024

Former-commit-id: 9147c258f6cd945fe337b0fe79b504a16b923802