Move overflow-y to gr-account-list internal list
The reply dialog set the max height of the gr-account-list component.
Because the autocomplete dropdown is positioned absolutely, it appeared
to not exist on the input as the account-list required more and more
scroll.

Applying a max-height and overflow-y to the list of account chips within
the account-list fixes the autocomplete dropdown placement.

Bug: Issue 7112
Change-Id: I5714f8f08c24d4ed2048b75dbc9ccf0e69b17774

Former-commit-id: a1ca71d0e5d672cbc480e6500d3cebb047faece4