Prevent autocomplete commit when input is unchanged
Normally, the enter key in the autocomplete commits to the current
suggestion. However, after a suggestion is committed and the suggestion
text remains in the input, pressing enter again will clear the
autocomplete since no suggestions are presented on account of no new
input being entered.

With this change, the enter key only fires a commit if the input text
has been changed since the last commit.

Bug: Issue 7196
Change-Id: I9e77e94461c99fdf218f51bbaee048b4b3793bec

Former-commit-id: a3b7747e4e3ff9648d72cd98542f9a144fd51648