Make cherrypick dialog 72ch wide
Previously, the cherrypick dialog's width was set to 30em, which was not
wide enough to display a reasonably sized commit message.

This change sets the width to 72ch, as that is a somewhat typical commit
message length, and definitely better than this was before.

The actualy length can be revisited when some of the commit message
editing conversations happen.

Bug: Issue 7183
Change-Id: I99d81fa98c2f7c7e87fc7d147de50b9476b076e6

Former-commit-id: 8b2d7817b61f97b26f7e290392fcc16be31f889d