DashboardsCollection: Don't serialize 'isDefault' as 'default'
The name 'default' can't be used as a member name because it's a
reserved word in Java. Instead it is defined as isDefault, and its
serialization overrided as "default" using gson's @SerializedName
annotation.

This is inconsistent with other REST APIs, which do not override
such names. For example "isPrivate" is allowed to be serialized
as "is_private".

Also, using @SerializedName means that if we want to move the
DashboardInfo class out to the extension API, we would also need
to introduce a dependency on gson in the extension API.

Remove the annotation and let it be serialized as "is_default".

Adjust the GWT UI to use the new name. Include support for the old
name for backwards compatibility in case the GWT UI and backend
are updated separately.

Change-Id: I3ddbb1f15a388c03386d088b9a45a7da1e060e0a

Former-commit-id: 8edc19969a7bef6824b7229b9a68eac0f2a4a637