MergeSuperSet: Remove indirect ChangeControl usage
ChangeControl was used in MergeSuperSet using an indirection inside
ChangeData. By calling changeData.changeControl(CurrentUser) ChangeData
would keep an instance of ChangeControl with the given user. This
instance was used inside MergeSuperSet only.

This commit removes this usage by removing the method on ChangeData that
lets callers specify a user for the ChangeControl instance and uses
ChangeControl inside MergeSuperSet only where really needed.

In addition, this commit moves visiblePatchSets from ChangeData to
ChangeControl to enforce callers to provide an explicit ChangeControl
instance to call this method on. However, this method is only needed
temporarily until drafts are deprecated.

Change-Id: I7f760f5d915bd268fe2f49120022114f3e246e45

Former-commit-id: 82071b883e00524a23cf2dca5436525c1bdb16be