ChangeIT: Add test for ignoring a change
When the Ignore endpoint was added, tests were missed.

Add a new test in ChangeIT, based on the existing test in AccountIT
which was added when support for ignoring a change by setting a
special star was added (which was before the Ignore endpoint).

Since we now have a specific API for ignoring the change, rename
the "ignoreChange" test in AccountIT to a more appropriate name.

Change-Id: I5b351af1c66c3721af50471e74a1302177c59437

Former-commit-id: 08cc35431aebcadb2ca4c0437bb0ff142c3ad55d