ListPlugins: Add convenience method to set all options from ListRequest
When invoking ListPlugins from the API implementation, it is necessary
to individually call each set method to set the necessary options.

To make this simpler, introduce a new method that sets them all at once,
using the builder pattern so that the caller can inline the whole thing
into a single line.

This will also have the advantage that if/when new options are added,
it will not be necessary to add extra code at the call location; it can
just be added in the new method.

Change-Id: Ifcece96c91bdb3350af2ba39e400150ad4328c85

Former-commit-id: bb3f66505ad9e28a6a18d89d701f4ab746532786