Fix docs URL construction in settings
When the docs base URL comes from different sources (probing or server
config) it may or may not include a slash. gr-settings-view's
_getFilterDocsLink method expected it to never end with a slash and
would thus construct invalid URLs when it did. With this change the
logic is fixed to normalize slashes and tests are added to cover the
various ways the method is called.

Bug: Issue 6860
Change-Id: I3ee16a292da2eee04eabad2f829175797998ae5a

Former-commit-id: 5693ea4b2b0c5f1de39616125f310020255e0fc4