Remove ChangeControl from ChangeData
This commit removes the ChangeControl member from ChangeData, removes
the getter method (changeControl()) as well as the additional create
statements in the Factory that are now obsolete.

This commit migrates all callers to construct a ChangeControl themselves
if needed or use other means to do their business.

This commit changes the behavior in terms of which user is used inside
aforementioned ChangeControls. Previously, calling
ChangeData#changeControl() would yield a ChangeControl scoped to the
change owner. By looking at the code, it must be assumed that callers
were not aware of this fact and it was an oversight to not scope the
control to the current user.
With this commit, all ChangeControls are scoped to the current user
instead, making checks upon them more meaningful. An exception to that
are methods inside ChangeData where the old behavior is kept for now
in lieu of a CurrentUser.

Change-Id: I80bf976e849e410dd9b5335dece6b55e7376de93

Former-commit-id: 72caf8f0f7adc7edf154ce77dfd4eec3f836a197